pyproject.toml
[project]
name = "image-search"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.5",
    "inference-sdk>=0.12.0",
    "matplotlib>=3.10.8",
    "notebook>=7.5.5",
    "open-clip-torch>=3.3.0",
    "opencv-python>=4.13.0.92",
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.56.0",
]
